fix: decimal days, storage symlink in Docker, dark mode on public pages
- Cast diffInDays to (int) in campaign card and show page
- Add storage:link symlink in Dockerfile (ln -sf)
- Add storage/app/public dir creation in Dockerfile
- Add dark mode init to public app layout

+++ src/resources/views/components/campaign-card.blade.php
@@ -7,7 +7,7 @@
     $paidInvoices = $campaign->invoices->where('is_paid', true);
     $donorCount = $paidInvoices->count();
     $recentDonors = $paidInvoices->take(3);
-    $daysLeft = $campaign->unlimited ? null : max(0, now()->diffInDays($campaign->posted_at->addDays($campaign->duration_days), false));
+    $daysLeft = $campaign->unlimited ? null : (int) max(0, now()->diffInDays($campaign->posted_at->addDays($campaign->duration_days), false));
 @endphp
 
 <a href="{{ route('campaign.show', $campaign) }}" class="block bg-white rounded-xl shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all duration-200 overflow-hidden">
